Proof of concept fix for keyboard on screen resize
This relies on the (currently unmegred) flutter/engine#9727. This WILL
NOT work until that has been rolled into flutter/flutter. This can be
tested locally by running this patch against a locally compiled
flutter/engine jar.

This uses the last cached input connection when the platform view is
locked, fixing an issue where the keyboard wasn't getting the right
input connection immediately after layout changes.

This code is very rough and should also be cleaned up before merging.
This is just a demo.